% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ugent.R
\name{theme_ugent}
\alias{theme_ugent}
\title{UGent-styled theme for ggplot2}
\usage{
theme_ugent(base_size = 20, faculty = NULL)
}
\arguments{
\item{base_size}{A numeric value determining the base font size for various text elements within the ggplot. This parameter sets a standard size from which relative sizes of other text elements (such as titles, subtitles, axis labels, and legend text) are calculated. The default value is set to 20.}

\item{faculty}{A character string representing the shortcode for a specific faculty within Ghent University. This parameter allows the theme customisation to reflect the unique identity of different faculties (e.g., "EB" for the Faculty of Economics and Business Administration) by applying their particular colour to some of the plot elements. If the provided shortcode is not recognised, or if \code{faculty} is left as \code{NULL}, the function defaults to using UGent yellow (\verb{#FFD200}) as the accent colour. For now, this parameter only influences the strip text's colour.}
}
\value{
A ggplot2 theme object.
}
\description{
This function applies a custom theme to ggplot objects.
The theme aims to create a visually appealing and consistent style for data visualisations,
adhering to UGent's branding guidelines.
It uses UGent Panno Text as the default font and
includes customised (colour) settings for plot titles, subtitles, captions, legends, axes, and panel grids.
}
\examples{
library(ggplot2)
library(ggugent)

# Using mtcars dataset as an example
ggplot(data = mtcars, mapping = aes(x = wt, y = mpg)) +
  geom_point() +
  geom_smooth(
    method = "lm",
    formula = y ~ x,
    alpha = .15,
    colour = "#1E64C8"
  ) +
  facet_wrap(~cyl, ncol = 3) +
  scale_x_continuous(limits = c(0, 5.5), breaks = seq(0, 5, 1)) +
  scale_y_continuous(limits = c(5, 35)) +
  labs(
    title = "Motor Trend car road tests",
    subtitle = "Vehicles miles per gallon vs weight by number of cylinders",
    caption = "Source: Motor Trend US magazine",
    x = "Weight (x1,000 lbs)",
    y = "Miles / (US) gallon"
  ) +
  theme_ugent(
    base_size = 20,
    faculty = "EB"
  )

}
